refactor(state): migrated Ship and Fleet definitions to separate types/shipTypes file
